extract client tick event to own class
